code formatting

+++ app/controllers/keep_it_ups_controller.rb
@@ -1,16 +1,12 @@
 class KeepItUpsController < ApplicationController
   def create
-    unless KeepItUp.exists?(user_id: current_user.id, tweet_id: params[:tweet_id])
-      KeepItUp.create(keep_it_up_params)
-    end
+    KeepItUp.create(keep_it_up_params) unless KeepItUp.exists?(user_id: current_user.id, tweet_id: params[:tweet_id])
     redirect_to tweet_path(params[:tweet_id])
   end
 
   def destroy
     keep_it_up = KeepItUp.find_by(user_id: current_user.id, tweet_id: params[:tweet_id])
-    if keep_it_up.present?
-      keep_it_up.destroy
-    end
+    keep_it_up.destroy if keep_it_up.present?
     redirect_to tweet_path(params[:tweet_id])
   end
